% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chronology}
\alias{chronology}
\title{\code{chronology} table}
\format{
A data frame (\code{tibble} object) with 254017 rows and 3 variables:
\describe{
\item{ID_MODEL}{int ...}
\item{ID_SAMPLE}{int ...}
\item{original_est_age}{int ...}
}
}
\usage{
chronology
}
\description{
Contains metadata for each chronology record.
}
\keyword{datasets}
