% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-publication.R
\name{get_publications}
\alias{get_publications}
\alias{get_publications.site}
\alias{get_publications.default}
\title{Get publications linked to a hydrological basin site}
\usage{
get_publications(.site_tbl, ...)

\method{get_publications}{site}(.site_tbl, ...)

\method{get_publications}{default}(.site_tbl, ...)
}
\arguments{
\item{.site_tbl}{\code{tibble} object.}

\item{...}{Optional parameters.}
}
\description{
Get publications linked to a hydrological basin site
}
\examples{
\dontrun{
conn <- open_conn_mysql("sys", "root")
get_publications(conn)
}
}
