% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_name}
\alias{model_name}
\title{\code{model_name} table}
\format{
A data frame (\code{tibble} object) with 18 rows and 2 variables:
\describe{
\item{ID_MODEL}{\code{Unsigned integer}: Unique identifier for the technique used to generate the age model}
\item{model_name}{\code{Text}: Name of the technique used to generate the age model}
}
}
\usage{
model_name
}
\description{
Contains metadata for each model name record.
}
\keyword{datasets}
