% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publication}
\alias{publication}
\title{\code{publication} table}
\format{
A data frame (\code{tibble} object) with 1163 rows and 4 variables:
\describe{
\item{ID_PUB}{\code{Unsigned integer}: Unique identifier for the publication}
\item{citation}{\code{Text}: The citation for the publication}
\item{pub_DOI_URL}{\code{Text}: The digital object identifier (doi) for the publication}
\item{bibentry}{\code{Text}: The publication citation in bibtex format}
}
}
\usage{
publication
}
\description{
Contains metadata for each publication used to retrieve data.
}
\keyword{datasets}
