% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample}
\alias{sample}
\title{\code{sample} table}
\format{
A data frame (\code{tibble} object) with 254763 rows and 7 variables:
\describe{
\item{ID_ENTITY}{int ...}
\item{ID_SAMPLE}{int ...}
\item{avg_depth}{int ...}
\item{sample_thickness}{int ...}
\item{charcoal_measurement}{int ...}
\item{analytical_sample_volume}{int ...}
}
}
\usage{
sample
}
\description{
Contains metadata for each sample record.
}
\keyword{datasets}
