% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site}
\alias{site}
\title{\code{site} table}
\format{
A data frame (\code{tibble} object) with 1480 rows and 12 variables:
\describe{
\item{ID_SITE}{\code{Unsigned integer}: Unique identifier for each site}
\item{site_name}{\code{Text}: Site name as given by original authors or as defined by us where there was no unique name given to the site}
\item{latitude}{\code{Double}: Latitude of the sampling site, given in decimal degrees, where N is positive and S is negative}
\item{longitude}{\code{Double}: Longitude of the sampling site in decimal degrees, where E is positive and W is negative}
\item{elevation}{\code{Double}: Elevation of the sampling site in metres above (+) or below (-) sea level}
\item{site_type}{\code{Text}: Information about type of site (e.g. lake, peatland, terrestrial)}
\item{water_depth}{\code{Double}: Water depth of the sampling site in metres}
\item{basin_size_class}{\code{Text}: Categorical estimate of basin size}
\item{catch_size_class}{\code{Text}: Categorical estimate of the hydrological catchment size}
\item{flow_type}{\code{Text}: Indication of whether there is inflow and/or outflow from the sampled site}
\item{basin_size_km2}{\code{Double}: Size of sampled site (e.g., lake or bog) in km2}
\item{catch_size_km2}{\code{Double}: Size of hydrological catchment in km2}
}
}
\usage{
site
}
\description{
Contains metadata for each site record.
}
\keyword{datasets}
