% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit}
\alias{unit}
\title{\code{unit} table}
\format{
A data frame (\code{tibble} object) with 202 rows and 2 variables:
\describe{
\item{ID_UNIT}{int ...}
\item{UNIT}{int ...}
}
}
\usage{
unit
}
\description{
Contains metadata for each unit record.
}
\keyword{datasets}
