% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_model}
\alias{age_model}
\title{\code{age_model} table}
\format{
A data frame (\code{tibble} object) with 121720 rows and 8 variables:
\describe{
\item{ID_MODEL}{int ...}
\item{ID_SAMPLE}{int ...}
\item{mean}{int ...}
\item{median}{int ...}
\item{UNCERT_5}{int ...}
\item{UNCERT_25}{int ...}
\item{UNCERT_75}{int ...}
\item{UNCERT_95}{int ...}
}
}
\usage{
age_model
}
\description{
Contains metadata for each age model record.
}
\keyword{datasets}
