% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chronology}
\alias{chronology}
\title{\code{chronology} table}
\format{
A data frame (\code{tibble} object) with 254017 rows and 3 variables:
\describe{
\item{ID_MODEL}{\code{Unsigned integer}: Unique identifier for the technique used to generate the original age model}
\item{ID_SAMPLE}{\code{Unsigned integer}: Unique identifier for the sample (as in \code{sample} table)}
\item{original_est_age}{\code{Integer}: Original estimated age}
}
}
\usage{
chronology
}
\description{
Contains metadata for each chronology record.
}
\keyword{datasets}
