% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_info}
\alias{date_info}
\title{\code{date_info} table}
\format{
A data frame (\code{tibble} object) with 16039 rows and 14 variables:
\describe{
\item{ID_ENTITY}{int ...}
\item{ID_DATE_INFO}{int ...}
\item{material_dated}{int ...}
\item{date_type}{int ...}
\item{avg_depth}{int ...}
\item{thickness}{int ...}
\item{lab_number}{int ...}
\item{age_C14}{int ...}
\item{age_calib}{int ...}
\item{error}{int ...}
\item{correlation_info}{int ...}
\item{age_used}{int ...}
\item{reason_age_not_used}{int ...}
\item{notes}{int ...}
}
}
\usage{
date_info
}
\description{
Contains metadata for each date info record.
}
\keyword{datasets}
