% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entity}
\alias{entity}
\title{\code{entity} table}
\format{
A data frame (\code{tibble} object) with 1682 rows and 13 variables:
\describe{
\item{ID_SITE}{int ...}
\item{ID_ENTITY}{int ...}
\item{entity_name}{int ...}
\item{latitude}{int ...}
\item{longitude}{int ...}
\item{elevation}{int ...}
\item{depositional_context}{int ...}
\item{measurement_method}{int ...}
\item{TYPE}{int ...}
\item{source}{int ...}
\item{core_location}{int ...}
\item{last_updated}{int ...}
\item{ID_UNIT}{int ...}
}
}
\usage{
entity
}
\description{
Contains metadata for each entity record.
}
\keyword{datasets}
