% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entity}
\alias{entity}
\title{Entity table}
\format{
A tibble with X rows and Y variables:
\describe{
  \item{A}{int ...}
}
}
\usage{
entity
}
\description{
Contains metadata for each entity
}
\keyword{datasets}
