% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entity_link_publication}
\alias{entity_link_publication}
\title{\code{entity_link_publication} table}
\format{
A data frame (\code{tibble} object) with 2147 rows and 2 variables:
\describe{
\item{ID_ENTITY}{\code{Unsigned integer}: Unique identifier for the entity (as in \code{entity} table)}
\item{ID_PUB}{\code{Unsigned integer}: Unique identifier for the publication (as in \code{publication} table)}
}
}
\usage{
entity_link_publication
}
\description{
Contains links between entities and publications.
}
\keyword{datasets}
