% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-site.R
\name{get_entities}
\alias{get_entities}
\alias{get_entities.site}
\alias{get_entities.default}
\alias{get_entities.tbl_df}
\title{Get entities linked to a hydrological basin site}
\usage{
get_entities(.data, ...)

\method{get_entities}{site}(.data, ...)

\method{get_entities}{default}(.data, ...)

\method{get_entities}{tbl_df}(.data, ...)
}
\arguments{
\item{.data}{\code{tibble} with site information.}

\item{...}{Optional parameters.}
}
\description{
Get entities linked to a hydrological basin site
}
\examples{
\dontrun{
conn <- open_conn_mysql("sys", "root")
get_entities(conn)
}
}
