% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publication}
\alias{publication}
\title{\code{publication} table}
\format{
A data frame (\code{tibble} object) with 1151 rows and 4 variables:
\describe{
\item{ID_PUB}{int ...}
\item{citation}{int ...}
\item{pub_DOI_URL}{int ...}
\item{bibentry}{int ...}
}
}
\usage{
publication
}
\description{
Contains metadata for each publication used to retrieve data.
}
\keyword{datasets}
