% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site}
\alias{site}
\title{Site table}
\format{
A tibble with X rows and Y variables:
\describe{
  \item{A}{int ...}
}
}
\usage{
site
}
\description{
Contains metadata for each site.
}
\keyword{datasets}
