% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site}
\alias{site}
\title{\code{site} table}
\format{
A data frame (\code{tibble} object) with 1477 rows and 12 variables:
\describe{
\item{ID_SITE}{int ...}
\item{site_name}{int ...}
\item{latitude}{int ...}
\item{longitude}{int ...}
\item{elevation}{int ...}
\item{site_type}{int ...}
\item{water_depth}{int ...}
\item{basin_size_class}{int ...}
\item{catch_size_class}{int ...}
\item{flow_type}{int ...}
\item{basin_size_km2}{int ...}
\item{catch_size_km2}{int ...}
}
}
\usage{
site
}
\description{
Contains metadata for each site record.
}
\keyword{datasets}
